Reorganise get Cypher query functions into individual files
