+++ src/neo4j/cypher-queries/award-ceremony/award-contextual-duplicate-record-count.js
@@ -1,0 +1,14 @@
+export default () => `
+	MATCH (ceremony:AwardCeremony { name: $name })<-[:PRESENTED_AT]-(award:Award { name: $award.name })
+		WHERE
+			(
+				$uuid IS NULL OR
+				$uuid <> ceremony.uuid
+			) AND
+			(
+				($award.differentiator IS NULL AND award.differentiator IS NULL) OR
+				$award.differentiator = award.differentiator
+			)
+
+	RETURN SIGN(COUNT(ceremony)) AS duplicateRecordCount
+`;
