+++ src/neo4j/cypher-queries/company/index.js
@@ -1,0 +1,5 @@
+import getShowQuery from './show';
+
+export {
+	getShowQuery
+};
