+++ src/neo4j/cypher-queries/material/index.js
@@ -1,0 +1,12 @@
+import { getCreateQuery, getUpdateQuery } from './create-update';
+import getEditQuery from './edit';
+import getListQuery from './list';
+import getShowQuery from './show';
+
+export {
+	getCreateQuery,
+	getEditQuery,
+	getUpdateQuery,
+	getShowQuery,
+	getListQuery
+};
