+++ src/neo4j/cypher-queries/material/list.js
@@ -1,0 +1,95 @@
+export default () => `
+	MATCH (material:Material)
+		WHERE NOT EXISTS((material)-[:HAS_SUB_MATERIAL]->(:Material))
+
+	OPTIONAL MATCH (material)-[entityRel:HAS_WRITING_ENTITY|USES_SOURCE_MATERIAL]->(entity)
+		WHERE entity:Person OR entity:Company OR entity:Material
+
+	OPTIONAL MATCH (entity:Material)-[sourceMaterialWriterRel:HAS_WRITING_ENTITY]->(sourceMaterialWriter)
+		WHERE sourceMaterialWriter:Person OR sourceMaterialWriter:Company
+
+	OPTIONAL MATCH (entity:Material)<-[:HAS_SUB_MATERIAL]-(entitySurMaterial:Material)
+
+	WITH material, entityRel, entity, entitySurMaterial, sourceMaterialWriterRel, sourceMaterialWriter
+		ORDER BY sourceMaterialWriterRel.creditPosition, sourceMaterialWriterRel.entityPosition
+
+	WITH
+		material,
+		entityRel,
+		entity,
+		entitySurMaterial,
+		sourceMaterialWriterRel.credit AS sourceMaterialWritingCreditName,
+		COLLECT(
+			CASE sourceMaterialWriter WHEN NULL
+				THEN null
+				ELSE sourceMaterialWriter { model: TOUPPER(HEAD(LABELS(sourceMaterialWriter))), .uuid, .name }
+			END
+		) AS sourceMaterialWriters
+
+	WITH material, entityRel, entity, entitySurMaterial,
+		COLLECT(
+			CASE SIZE(sourceMaterialWriters) WHEN 0
+				THEN null
+				ELSE {
+					model: 'WRITING_CREDIT',
+					name: COALESCE(sourceMaterialWritingCreditName, 'by'),
+					entities: sourceMaterialWriters
+				}
+			END
+		) AS sourceMaterialWritingCredits
+		ORDER BY entityRel.creditPosition, entityRel.entityPosition
+
+	WITH material, entityRel.credit AS writingCreditName,
+		COLLECT(
+			CASE entity WHEN NULL
+				THEN null
+				ELSE entity {
+					model: TOUPPER(HEAD(LABELS(entity))),
+					.uuid,
+					.name,
+					.format,
+					.year,
+					surMaterial: CASE entitySurMaterial WHEN NULL
+						THEN null
+						ELSE entitySurMaterial { model: 'MATERIAL', .uuid, .name }
+					END,
+					writingCredits: sourceMaterialWritingCredits
+				}
+			END
+		) AS entities
+
+	WITH material, writingCreditName,
+		[entity IN entities | CASE entity.model WHEN 'MATERIAL'
+			THEN entity
+			ELSE entity { .model, .uuid, .name }
+		END] AS entities
+
+	WITH material,
+		COLLECT(
+			CASE SIZE(entities) WHEN 0
+				THEN null
+				ELSE {
+					model: 'WRITING_CREDIT',
+					name: COALESCE(writingCreditName, 'by'),
+					entities: entities
+				}
+			END
+		) AS writingCredits
+
+	OPTIONAL MATCH (material)<-[:HAS_SUB_MATERIAL]-(surMaterial:Material)
+
+	RETURN
+		'MATERIAL' AS model,
+		material.uuid AS uuid,
+		material.name AS name,
+		material.format AS format,
+		material.year AS year,
+		CASE surMaterial WHEN NULL
+			THEN null
+			ELSE surMaterial { model: 'MATERIAL', .uuid, .name }
+		END AS surMaterial,
+		writingCredits
+		ORDER BY material.year DESC, material.name
+
+	LIMIT 100
+`;
