+++ src/neo4j/cypher-queries/material/show.js
@@ -1,333 +1,4 @@
-import { ACTIONS } from '../../utils/constants';
-
-const getCreateUpdateQuery = action => {
-
-	const createUpdateQueryOpeningMap = {
-		[ACTIONS.CREATE]: `
-			CREATE (material:Material {
-				uuid: $uuid,
-				name: $name,
-				differentiator: $differentiator,
-				format: $format,
-				year: $year
-			})
-		`,
-		[ACTIONS.UPDATE]: `
-			MATCH (material:Material { uuid: $uuid })
-
-			OPTIONAL MATCH (material)-[originalVersionMaterialRel:SUBSEQUENT_VERSION_OF]->(:Material)
-
-			DELETE originalVersionMaterialRel
-
-			WITH DISTINCT material
-
-			OPTIONAL MATCH (material)-[writerRel:HAS_WRITING_ENTITY]->(entity)
-				WHERE entity:Person OR entity:Company
-
-			DELETE writerRel
-
-			WITH DISTINCT material
-
-			OPTIONAL MATCH (material)-[sourceMaterialCreditRel:USES_SOURCE_MATERIAL]->(:Material)
-
-			DELETE sourceMaterialCreditRel
-
-			WITH DISTINCT material
-
-			OPTIONAL MATCH (material)-[subMaterialRel:HAS_SUB_MATERIAL]->(:Material)
-
-			DELETE subMaterialRel
-
-			WITH DISTINCT material
-
-			OPTIONAL MATCH (material)-[characterRel:DEPICTS]->(:Character)
-
-			DELETE characterRel
-
-			WITH DISTINCT material
-
-			SET
-				material.name = $name,
-				material.differentiator = $differentiator,
-				material.format = $format,
-				material.year = $year
-		`
-	};
-
-	return `
-		${createUpdateQueryOpeningMap[action]}
-
-		WITH material
-
-		OPTIONAL MATCH (existingOriginalVersionMaterial:Material { name: $originalVersionMaterial.name })
-			WHERE
-				(
-					$originalVersionMaterial.differentiator IS NULL AND
-					existingOriginalVersionMaterial.differentiator IS NULL
-				) OR
-				$originalVersionMaterial.differentiator = existingOriginalVersionMaterial.differentiator
-
-		FOREACH (item IN CASE $originalVersionMaterial.name WHEN NULL THEN [] ELSE [1] END |
-			MERGE (originalVersionMaterial:Material {
-				uuid: COALESCE(existingOriginalVersionMaterial.uuid, $originalVersionMaterial.uuid),
-				name: $originalVersionMaterial.name
-			})
-				ON CREATE SET originalVersionMaterial.differentiator = $originalVersionMaterial.differentiator
-
-			CREATE (material)-[:SUBSEQUENT_VERSION_OF]->(originalVersionMaterial)
-		)
-
-		WITH material
-
-		UNWIND (CASE $writingCredits WHEN [] THEN [{ entities: [] }] ELSE $writingCredits END) AS writingCredit
-
-			UNWIND
-				CASE SIZE([entity IN writingCredit.entities WHERE entity.model = 'PERSON']) WHEN 0
-					THEN [null]
-					ELSE [entity IN writingCredit.entities WHERE entity.model = 'PERSON']
-				END AS writingPersonParam
-
-				OPTIONAL MATCH (existingWritingPerson:Person { name: writingPersonParam.name })
-					WHERE
-						(writingPersonParam.differentiator IS NULL AND existingWritingPerson.differentiator IS NULL) OR
-						writingPersonParam.differentiator = existingWritingPerson.differentiator
-
-				FOREACH (item IN CASE writingPersonParam WHEN NULL THEN [] ELSE [1] END |
-					MERGE (writingPerson:Person {
-						uuid: COALESCE(existingWritingPerson.uuid, writingPersonParam.uuid),
-						name: writingPersonParam.name
-					})
-						ON CREATE SET writingPerson.differentiator = writingPersonParam.differentiator
-
-					CREATE (material)-
-						[:HAS_WRITING_ENTITY {
-							creditPosition: writingCredit.position,
-							entityPosition: writingPersonParam.position,
-							credit: writingCredit.name,
-							creditType: writingCredit.creditType
-						}]->(writingPerson)
-				)
-
-			WITH DISTINCT material, writingCredit
-
-			UNWIND
-				CASE SIZE([entity IN writingCredit.entities WHERE entity.model = 'COMPANY']) WHEN 0
-					THEN [null]
-					ELSE [entity IN writingCredit.entities WHERE entity.model = 'COMPANY']
-				END AS writingCompanyParam
-
-				OPTIONAL MATCH (existingWritingCompany:Company { name: writingCompanyParam.name })
-					WHERE
-						(
-							writingCompanyParam.differentiator IS NULL AND
-							existingWritingCompany.differentiator IS NULL
-						) OR
-						writingCompanyParam.differentiator = existingWritingCompany.differentiator
-
-				FOREACH (item IN CASE writingCompanyParam WHEN NULL THEN [] ELSE [1] END |
-					MERGE (writingCompany:Company {
-						uuid: COALESCE(existingWritingCompany.uuid, writingCompanyParam.uuid),
-						name: writingCompanyParam.name
-					})
-						ON CREATE SET writingCompany.differentiator = writingCompanyParam.differentiator
-
-					CREATE (material)-
-						[:HAS_WRITING_ENTITY {
-							creditPosition: writingCredit.position,
-							entityPosition: writingCompanyParam.position,
-							credit: writingCredit.name,
-							creditType: writingCredit.creditType
-						}]->(writingCompany)
-				)
-
-			WITH DISTINCT material, writingCredit
-
-			UNWIND
-				CASE SIZE([entity IN writingCredit.entities WHERE entity.model = 'MATERIAL']) WHEN 0
-					THEN [null]
-					ELSE [entity IN writingCredit.entities WHERE entity.model = 'MATERIAL']
-				END AS sourceMaterialParam
-
-				OPTIONAL MATCH (existingSourceMaterial:Material { name: sourceMaterialParam.name })
-					WHERE
-						(
-							sourceMaterialParam.differentiator IS NULL AND
-							existingSourceMaterial.differentiator IS NULL
-						) OR
-						sourceMaterialParam.differentiator = existingSourceMaterial.differentiator
-
-				FOREACH (item IN CASE sourceMaterialParam WHEN NULL THEN [] ELSE [1] END |
-					MERGE (sourceMaterial:Material {
-						uuid: COALESCE(existingSourceMaterial.uuid, sourceMaterialParam.uuid),
-						name: sourceMaterialParam.name
-					})
-						ON CREATE SET sourceMaterial.differentiator = sourceMaterialParam.differentiator
-
-					CREATE (material)-
-						[:USES_SOURCE_MATERIAL {
-							creditPosition: writingCredit.position,
-							entityPosition: sourceMaterialParam.position,
-							credit: writingCredit.name
-						}]->(sourceMaterial)
-				)
-
-		WITH DISTINCT material
-
-			UNWIND (CASE $subMaterials WHEN [] THEN [null] ELSE $subMaterials END) AS subMaterialParam
-
-				OPTIONAL MATCH (existingSubMaterial:Material { name: subMaterialParam.name })
-					WHERE
-						(subMaterialParam.differentiator IS NULL AND existingSubMaterial.differentiator IS NULL) OR
-						subMaterialParam.differentiator = existingSubMaterial.differentiator
-
-				FOREACH (item IN CASE subMaterialParam WHEN NULL THEN [] ELSE [1] END |
-					MERGE (subMaterial:Material {
-						uuid: COALESCE(existingSubMaterial.uuid, subMaterialParam.uuid),
-						name: subMaterialParam.name
-					})
-						ON CREATE SET
-							subMaterial.differentiator = subMaterialParam.differentiator,
-							subMaterial.format = subMaterialParam.format,
-							subMaterial.year = subMaterialParam.year
-
-					CREATE (material)-[:HAS_SUB_MATERIAL { position: subMaterialParam.position }]->(subMaterial)
-				)
-
-		WITH DISTINCT material
-
-		UNWIND (CASE $characterGroups WHEN [] THEN [{ characters: [] }] ELSE $characterGroups END) AS characterGroup
-
-			UNWIND (CASE characterGroup.characters WHEN []
-				THEN [null]
-				ELSE characterGroup.characters
-			END) AS characterParam
-
-				OPTIONAL MATCH (existingCharacter:Character {
-					name: COALESCE(characterParam.underlyingName, characterParam.name)
-				})
-					WHERE
-						(characterParam.differentiator IS NULL AND existingCharacter.differentiator IS NULL) OR
-						characterParam.differentiator = existingCharacter.differentiator
-
-				FOREACH (item IN CASE characterParam WHEN NULL THEN [] ELSE [1] END |
-					MERGE (character:Character {
-						uuid: COALESCE(existingCharacter.uuid, characterParam.uuid),
-						name: COALESCE(characterParam.underlyingName, characterParam.name)
-					})
-						ON CREATE SET character.differentiator = characterParam.differentiator
-
-					CREATE (material)-
-						[:DEPICTS {
-							groupPosition: characterGroup.position,
-							characterPosition: characterParam.position,
-							displayName: CASE characterParam.underlyingName WHEN NULL
-								THEN null
-								ELSE characterParam.name
-							END,
-							qualifier: characterParam.qualifier,
-							group: characterGroup.name
-						}]->(character)
-				)
-
-		WITH DISTINCT material
-
-		${getEditQuery()}
-	`;
-
-};
-
-const getCreateQuery = () => getCreateUpdateQuery(ACTIONS.CREATE);
-
-const getEditQuery = () => `
-	MATCH (material:Material { uuid: $uuid })
-
-	OPTIONAL MATCH (material)-[:SUBSEQUENT_VERSION_OF]->(originalVersionMaterial:Material)
-
-	OPTIONAL MATCH (material)-[entityRel:HAS_WRITING_ENTITY|USES_SOURCE_MATERIAL]->(entity)
-		WHERE entity:Person OR entity:Company OR entity:Material
-
-	WITH material, originalVersionMaterial, entityRel, entity
-		ORDER BY entityRel.creditPosition, entityRel.entityPosition
-
-	WITH
-		material,
-		originalVersionMaterial,
-		entityRel.credit AS writingCreditName,
-		entityRel.creditType AS writingCreditType,
-		COLLECT(
-			CASE entity WHEN NULL
-				THEN null
-				ELSE entity { model: TOUPPER(HEAD(LABELS(entity))), .name, .differentiator }
-			END
-		) + [{}] AS entities
-
-	WITH material, originalVersionMaterial,
-		COLLECT(
-			CASE WHEN writingCreditName IS NULL AND SIZE(entities) = 1
-				THEN null
-				ELSE {
-					name: writingCreditName,
-					creditType: writingCreditType,
-					entities: entities
-				}
-			END
-		) + [{ entities: [{}] }] AS writingCredits
-
-	OPTIONAL MATCH (material)-[subMaterialRel:HAS_SUB_MATERIAL]->(subMaterial:Material)
-
-	WITH material, originalVersionMaterial, writingCredits, subMaterialRel, subMaterial
-		ORDER BY subMaterialRel.position
-
-	WITH material, originalVersionMaterial, writingCredits,
-		COLLECT(
-			CASE subMaterial WHEN NULL
-				THEN null
-				ELSE subMaterial { .name, .differentiator, .format, .year }
-			END
-		) + [{}] AS subMaterials
-
-	OPTIONAL MATCH (material)-[characterRel:DEPICTS]->(character:Character)
-
-	WITH material, originalVersionMaterial, writingCredits, subMaterials, characterRel, character
-		ORDER BY characterRel.groupPosition, characterRel.characterPosition
-
-	WITH material, originalVersionMaterial, writingCredits, subMaterials, characterRel.group AS characterGroupName,
-		COLLECT(
-			CASE character WHEN NULL
-				THEN null
-				ELSE character {
-					name: COALESCE(characterRel.displayName, character.name),
-					underlyingName: CASE characterRel.displayName WHEN NULL THEN null ELSE character.name END,
-					.differentiator,
-					qualifier: characterRel.qualifier,
-					group: characterRel.group
-				}
-			END
-		) + [{}] AS characters
-
-	RETURN
-		material.uuid AS uuid,
-		material.name AS name,
-		material.differentiator AS differentiator,
-		material.format AS format,
-		material.year AS year,
-		{
-			name: COALESCE(originalVersionMaterial.name, ''),
-			differentiator: COALESCE(originalVersionMaterial.differentiator, '')
-		} AS originalVersionMaterial,
-		writingCredits,
-		subMaterials,
-		COLLECT(
-			CASE WHEN characterGroupName IS NULL AND SIZE(characters) = 1
-				THEN null
-				ELSE { name: characterGroupName, characters: characters }
-			END
-		) + [{ characters: [{}] }] AS characterGroups
-`;
-
-const getUpdateQuery = () => getCreateUpdateQuery(ACTIONS.UPDATE);
-
-const getShowQuery = () => `
+export default () => `
 	MATCH (material:Material { uuid: $uuid })
 
 	OPTIONAL MATCH (material)-[:SUBSEQUENT_VERSION_OF]-(originalOrSubsequentVersionMaterial)
@@ -1698,108 +1369,4 @@ const getShowQuery = () => `
 		awards,
 		subsequentVersionMaterialAwards,
 		COLLECT(sourcingMaterialAward { model: 'AWARD', .uuid, .name, ceremonies }) AS sourcingMaterialAwards
-	`;
-
-const getListQuery = () => `
-	MATCH (material:Material)
-		WHERE NOT EXISTS((material)-[:HAS_SUB_MATERIAL]->(:Material))
-
-	OPTIONAL MATCH (material)-[entityRel:HAS_WRITING_ENTITY|USES_SOURCE_MATERIAL]->(entity)
-		WHERE entity:Person OR entity:Company OR entity:Material
-
-	OPTIONAL MATCH (entity:Material)-[sourceMaterialWriterRel:HAS_WRITING_ENTITY]->(sourceMaterialWriter)
-		WHERE sourceMaterialWriter:Person OR sourceMaterialWriter:Company
-
-	OPTIONAL MATCH (entity:Material)<-[:HAS_SUB_MATERIAL]-(entitySurMaterial:Material)
-
-	WITH material, entityRel, entity, entitySurMaterial, sourceMaterialWriterRel, sourceMaterialWriter
-		ORDER BY sourceMaterialWriterRel.creditPosition, sourceMaterialWriterRel.entityPosition
-
-	WITH
-		material,
-		entityRel,
-		entity,
-		entitySurMaterial,
-		sourceMaterialWriterRel.credit AS sourceMaterialWritingCreditName,
-		COLLECT(
-			CASE sourceMaterialWriter WHEN NULL
-				THEN null
-				ELSE sourceMaterialWriter { model: TOUPPER(HEAD(LABELS(sourceMaterialWriter))), .uuid, .name }
-			END
-		) AS sourceMaterialWriters
-
-	WITH material, entityRel, entity, entitySurMaterial,
-		COLLECT(
-			CASE SIZE(sourceMaterialWriters) WHEN 0
-				THEN null
-				ELSE {
-					model: 'WRITING_CREDIT',
-					name: COALESCE(sourceMaterialWritingCreditName, 'by'),
-					entities: sourceMaterialWriters
-				}
-			END
-		) AS sourceMaterialWritingCredits
-		ORDER BY entityRel.creditPosition, entityRel.entityPosition
-
-	WITH material, entityRel.credit AS writingCreditName,
-		COLLECT(
-			CASE entity WHEN NULL
-				THEN null
-				ELSE entity {
-					model: TOUPPER(HEAD(LABELS(entity))),
-					.uuid,
-					.name,
-					.format,
-					.year,
-					surMaterial: CASE entitySurMaterial WHEN NULL
-						THEN null
-						ELSE entitySurMaterial { model: 'MATERIAL', .uuid, .name }
-					END,
-					writingCredits: sourceMaterialWritingCredits
-				}
-			END
-		) AS entities
-
-	WITH material, writingCreditName,
-		[entity IN entities | CASE entity.model WHEN 'MATERIAL'
-			THEN entity
-			ELSE entity { .model, .uuid, .name }
-		END] AS entities
-
-	WITH material,
-		COLLECT(
-			CASE SIZE(entities) WHEN 0
-				THEN null
-				ELSE {
-					model: 'WRITING_CREDIT',
-					name: COALESCE(writingCreditName, 'by'),
-					entities: entities
-				}
-			END
-		) AS writingCredits
-
-	OPTIONAL MATCH (material)<-[:HAS_SUB_MATERIAL]-(surMaterial:Material)
-
-	RETURN
-		'MATERIAL' AS model,
-		material.uuid AS uuid,
-		material.name AS name,
-		material.format AS format,
-		material.year AS year,
-		CASE surMaterial WHEN NULL
-			THEN null
-			ELSE surMaterial { model: 'MATERIAL', .uuid, .name }
-		END AS surMaterial,
-		writingCredits
-		ORDER BY material.year DESC, material.name
-
-	LIMIT 100
 `;
-
-export {
-	getCreateQuery,
-	getEditQuery,
-	getUpdateQuery,
-	getShowQuery,
-	getListQuery
-};
