+++ src/neo4j/cypher-queries/person/index.js
@@ -1,0 +1,5 @@
+import getShowQuery from './show';
+
+export {
+	getShowQuery
+};
