+++ src/neo4j/cypher-queries/shared/create.js
@@ -1,0 +1,10 @@
+import { getEditQuery } from '.';
+import { MODEL_TO_NODE_LABEL_MAP } from '../../../utils/constants';
+
+export default model => `
+	CREATE (n:${MODEL_TO_NODE_LABEL_MAP[model]} { uuid: $uuid, name: $name, differentiator: $differentiator })
+
+	WITH n
+
+	${getEditQuery(model)}
+`;
