+++ src/neo4j/cypher-queries/shared/duplicate-record-count.js
@@ -1,0 +1,16 @@
+import { MODEL_TO_NODE_LABEL_MAP } from '../../../utils/constants';
+
+export default model => `
+	MATCH (n:${MODEL_TO_NODE_LABEL_MAP[model]} { name: $name })
+		WHERE
+			(
+				($differentiator IS NULL AND n.differentiator IS NULL) OR
+				$differentiator = n.differentiator
+			) AND
+			(
+				$uuid IS NULL OR
+				$uuid <> n.uuid
+			)
+
+	RETURN SIGN(COUNT(n)) AS duplicateRecordCount
+`;
