+++ src/neo4j/cypher-queries/shared/existence.js
@@ -1,0 +1,7 @@
+import { MODEL_TO_NODE_LABEL_MAP } from '../../../utils/constants';
+
+export default model => `
+	MATCH (n:${MODEL_TO_NODE_LABEL_MAP[model]} { uuid: $uuid })
+
+	RETURN n
+`;
