+++ src/neo4j/cypher-queries/shared/index.js
@@ -1,0 +1,17 @@
+import getCreateQuery from './create';
+import getDeleteQuery from './delete';
+import getDuplicateRecordCountQuery from './duplicate-record-count';
+import getEditQuery from './edit';
+import getExistenceQuery from './existence';
+import getListQuery from './list';
+import getUpdateQuery from './update';
+
+export {
+	getExistenceQuery,
+	getDuplicateRecordCountQuery,
+	getCreateQuery,
+	getEditQuery,
+	getUpdateQuery,
+	getDeleteQuery,
+	getListQuery
+};
