+++ src/neo4j/cypher-queries/shared/list.js
@@ -1,0 +1,14 @@
+import { MODEL_TO_NODE_LABEL_MAP } from '../../../utils/constants';
+
+export default model => `
+	MATCH (n:${MODEL_TO_NODE_LABEL_MAP[model]})
+
+	RETURN
+		'${model}' AS model,
+		n.uuid AS uuid,
+		n.name AS name
+
+	ORDER BY n.name
+
+	LIMIT 100
+`;
